HFUTProject1/requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.18.2
transformers==4.31.0
onnxruntime==1.15.1
fastapi==0.99.1
uvicorn==0.22.0
python-multipart==0.0.6
scikit-learn==1.3.0
einops==0.6.1
tensorboard==2.13.0
pyyaml==6.0.1